refactor(types): swap in-package imports to rllm.types
Migrate 41 active in-package files from `from rllm.agents.agent
import …` to `from rllm.types import …` for the trajectory data
types (Episode, Step, Trajectory, Action, TrajectoryGroup,
_DEFAULT_TRAJ_NAME). Three files (rllm/__init__.py,
rllm/workflows/workflow.py, rllm/workflows/simple_workflow.py)
keep a separate `from rllm.agents.agent import BaseAgent` line —
the BaseAgent ABC stays in rllm.agents.agent for now and will
move to rllm/v0/ alongside the AgentExecutionEngine path.

Class identity is preserved by the re-export shim so this is a
purely cosmetic move; tests pass with zero new failures.

Untouched: rllm/agents/*_agent.py, rllm/engine/agent_execution_engine.py,
rllm/environments/*, rllm/trainer/deprecated/* — all on the legacy
path slated for the v0/ relocation.
