+++ rllm/engine/agent_sdk_engine.py
@@ -17,7 +17,6 @@ try:
 except ImportError as err:
     raise ImportError("AgentSdkEngine requires extra dependencies. Install with: pip install rllm[train]") from err
 
-from rllm.agents.agent import Episode, Trajectory
 from rllm.engine.rollout import ModelOutput, RolloutEngine
 from rllm.engine.rollout.verl_engine import VerlEngine
 from rllm.sdk.data_process import group_steps, trace_to_step
@@ -26,6 +25,7 @@ from rllm.sdk.proxy.proxy_manager import VerlProxyManager
 from rllm.sdk.session import SESSION_BACKEND
 from rllm.sdk.session.base import wrap_with_session_context
 from rllm.sdk.store.sqlite_store import SqliteTraceStore
+from rllm.types import Episode, Trajectory
 from rllm.types import Trajectory as BaseTrajectory
 from rllm.utils import colorful_print
 from rllm.workflows.workflow import TerminationReason
