+++ rllm/engine/agent_workflow_engine.py
@@ -10,8 +10,8 @@ from typing import TYPE_CHECKING
 import numpy as np
 from tqdm import tqdm
 
-from rllm.agents.agent import Episode
 from rllm.engine.rollout import RolloutEngine
+from rllm.types import Episode
 from rllm.utils import colorful_print
 from rllm.workflows.workflow import TerminationReason, Workflow
