+++ rllm/experimental/common/rejection_sampling.py
@@ -7,8 +7,8 @@ tracking episode-level correctness metrics while filtering based on reward varia
 
 from dataclasses import dataclass, field
 
-from rllm.agents.agent import Episode, TrajectoryGroup
 from rllm.experimental.common.config import RejectionSamplingConfig
+from rllm.types import Episode, TrajectoryGroup
 
 
 @dataclass
