+++ rllm/experimental/engine/unified_workflow_engine.py
@@ -10,8 +10,8 @@ from typing import TYPE_CHECKING, Any
 
 from tqdm import tqdm
 
-from rllm.agents.agent import Episode
 from rllm.experimental.rollout import RolloutEngine
+from rllm.types import Episode
 from rllm.workflows.store import Store
 from rllm.workflows.workflow import TerminationReason, Workflow
