+++ rllm/experimental/protocol.py
@@ -14,10 +14,10 @@ from typing import TYPE_CHECKING, Generic, TypeVar
 
 from omegaconf import DictConfig
 
-from rllm.agents.agent import Episode
 from rllm.data import Dataset
 from rllm.experimental.common.advantage import AlgorithmConfig, collect_reward_and_advantage_from_trajectory_groups
 from rllm.experimental.rollout import RolloutEngine
+from rllm.types import Episode
 
 if TYPE_CHECKING:
     from rllm.experimental.engine.unified_workflow_engine import UnifiedWorkflowEngine
