+++ rllm/workflows/workflow.py
@@ -8,9 +8,10 @@ from functools import partial
 
 import numpy as np
 
-from rllm.agents.agent import BaseAgent, Episode, Trajectory
+from rllm.agents.agent import BaseAgent
 from rllm.engine.rollout.rollout_engine import RolloutEngine
 from rllm.environments.base.base_env import BaseEnv
+from rllm.types import Episode, Trajectory
 from rllm.workflows.store import Store
